Refactor FoundationContent and MandatoryConfigurationsContent for improved accessibility and error handling
